fix: SYSTEM ERROR reported after analysis completed
The error happens in the chord callback which is triggerred after the
workflow execution has completed, to finalize the ingestion process
for ingestion workflows. As the specific case of other workflows is new
to the WES, the flag was set to True in the
case of an analysis or an export workflow.